chore(index): rename main.js to index.js
